app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
clickhouse-driver==0.2.3
discord.py @ git+https://github.com/Rapptz/discord.py@1a90327202363ddda57df290cd8db1cab5d4832a
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
tzdata==2022.1
tzlocal==4.2
yarl==1.7.2
